Django-Middleware-0x03/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0
drf-nested-routers>=0.95,<1.0
django-environ>=0.10,<1.0
djangorestframework-simplejwt>=5.0,<6.0